test/smoke/requirements.txt
certifi==2019.11.28
chardet==3.0.4
coverage==5.0.3
deepdiff==4.2.0
idna==2.8
ordered-set==3.1.1
python-bitcointx==1.1.1.post2
python-dashtx==0.1.0
web3==7.5.0
rlp==4.1.0
requests==2.32.3
six==1.14.0
tenacity==6.1.0
urllib3==1.25.10
tqdm==4.45.0
websocket-client==0.57.0
bech32==1.2.0
terra-sdk==2.0.4
terra-proto==1.0.1
durationpy==0.5
responses==0.15.0
radix-engine-toolkit==1.0.12
jsonpickle==4.0.2
